Break the fuzzy matcher out into its own file.
